+++ src/celt/mdct.rs
@@ -34,47 +34,45 @@ pub unsafe fn clt_mdct_forward_c(
     scale = st.scale;
     let trig = l.trig[shift as usize];
     N = l.n >> shift;
-    N2 = N >> 1 as i32;
-    N4 = N >> 2 as i32;
+    N2 = N >> 1;
+    N4 = N >> 2;
     let vla = N2 as usize;
     let mut f: Vec<f32> = ::std::vec::from_elem(0., vla);
     let vla_0 = N4 as usize;
     let mut f2: Vec<kiss_fft_cpx> = ::std::vec::from_elem(kiss_fft_cpx::zero(), vla_0);
-    let mut xp1: *const f32 = in_0.offset((overlap >> 1 as i32) as isize);
+    let mut xp1: *const f32 = in_0.offset((overlap >> 1) as isize);
     let mut xp2: *const f32 = in_0
         .offset(N2 as isize)
-        .offset(-(1 as i32 as isize))
-        .offset((overlap >> 1 as i32) as isize);
+        .offset(-(1 as isize))
+        .offset((overlap >> 1) as isize);
     let mut yp: *mut f32 = f.as_mut_ptr();
-    let mut wp1: *const opus_val16 = window.offset((overlap >> 1 as i32) as isize);
-    let mut wp2: *const opus_val16 = window
-        .offset((overlap >> 1 as i32) as isize)
-        .offset(-(1 as i32 as isize));
-    i = 0 as i32;
-    while i < overlap + 3 as i32 >> 2 as i32 {
+    let mut wp1: *const opus_val16 = window.offset((overlap >> 1) as isize);
+    let mut wp2: *const opus_val16 = window.offset((overlap >> 1) as isize).offset(-(1 as isize));
+    i = 0;
+    while i < overlap + 3 >> 2 {
         let fresh0 = yp;
         yp = yp.offset(1);
         *fresh0 = *wp2 * *xp1.offset(N2 as isize) + *wp1 * *xp2;
         let fresh1 = yp;
         yp = yp.offset(1);
         *fresh1 = *wp1 * *xp1 - *wp2 * *xp2.offset(-N2 as isize);
-        xp1 = xp1.offset(2 as i32 as isize);
-        xp2 = xp2.offset(-(2 as i32 as isize));
-        wp1 = wp1.offset(2 as i32 as isize);
-        wp2 = wp2.offset(-(2 as i32 as isize));
+        xp1 = xp1.offset(2 as isize);
+        xp2 = xp2.offset(-(2 as isize));
+        wp1 = wp1.offset(2 as isize);
+        wp2 = wp2.offset(-(2 as isize));
         i += 1;
     }
     wp1 = window;
-    wp2 = window.offset(overlap as isize).offset(-(1 as i32 as isize));
-    while i < N4 - (overlap + 3 as i32 >> 2 as i32) {
+    wp2 = window.offset(overlap as isize).offset(-(1 as isize));
+    while i < N4 - (overlap + 3 >> 2) {
         let fresh2 = yp;
         yp = yp.offset(1);
         *fresh2 = *xp2;
         let fresh3 = yp;
         yp = yp.offset(1);
         *fresh3 = *xp1;
-        xp1 = xp1.offset(2 as i32 as isize);
-        xp2 = xp2.offset(-(2 as i32 as isize));
+        xp1 = xp1.offset(2 as isize);
+        xp2 = xp2.offset(-(2 as isize));
         i += 1;
     }
     while i < N4 {
@@ -84,15 +82,15 @@ pub unsafe fn clt_mdct_forward_c(
         let fresh5 = yp;
         yp = yp.offset(1);
         *fresh5 = *wp2 * *xp1 + *wp1 * *xp2.offset(N2 as isize);
-        xp1 = xp1.offset(2 as i32 as isize);
-        xp2 = xp2.offset(-(2 as i32 as isize));
-        wp1 = wp1.offset(2 as i32 as isize);
-        wp2 = wp2.offset(-(2 as i32 as isize));
+        xp1 = xp1.offset(2 as isize);
+        xp2 = xp2.offset(-(2 as isize));
+        wp1 = wp1.offset(2 as isize);
+        wp2 = wp2.offset(-(2 as isize));
         i += 1;
     }
     let mut yp_0: *mut f32 = f.as_mut_ptr();
     let t = trig;
-    i = 0 as i32;
+    i = 0;
     while i < N4 {
         let mut yc: kiss_fft_cpx = kiss_fft_cpx::zero();
         let mut t0: f32 = 0.;
@@ -121,9 +119,9 @@ pub unsafe fn clt_mdct_forward_c(
     opus_fft_impl(st, &mut f2);
     let mut fp: *const kiss_fft_cpx = f2.as_mut_ptr();
     let mut yp1: *mut f32 = out;
-    let mut yp2: *mut f32 = out.offset((stride * (N2 - 1 as i32)) as isize);
+    let mut yp2: *mut f32 = out.offset((stride * (N2 - 1)) as isize);
     let t = trig;
-    i = 0 as i32;
+    i = 0;
     while i < N4 {
         let mut yr_0: f32 = 0.;
         let mut yi_0: f32 = 0.;
@@ -132,8 +130,8 @@ pub unsafe fn clt_mdct_forward_c(
         *yp1 = yr_0;
         *yp2 = yi_0;
         fp = fp.offset(1);
-        yp1 = yp1.offset((2 as i32 * stride) as isize);
-        yp2 = yp2.offset(-((2 as i32 * stride) as isize));
+        yp1 = yp1.offset((2 * stride) as isize);
+        yp2 = yp2.offset(-((2 * stride) as isize));
         i += 1;
     }
 }
@@ -153,14 +151,14 @@ pub unsafe fn clt_mdct_backward_c(
     let mut N4: i32 = 0;
     let trig = l.trig[shift as usize];
     N = l.n >> shift;
-    N2 = N >> 1 as i32;
-    N4 = N >> 2 as i32;
+    N2 = N >> 1;
+    N4 = N >> 2;
     let mut xp1: *const f32 = in_0;
-    let mut xp2: *const f32 = in_0.offset((stride * (N2 - 1 as i32)) as isize);
-    let yp: *mut f32 = out.offset((overlap >> 1 as i32) as isize);
+    let mut xp2: *const f32 = in_0.offset((stride * (N2 - 1)) as isize);
+    let yp: *mut f32 = out.offset((overlap >> 1) as isize);
     let t = trig;
     let mut bitrev: *const i16 = l.kfft[shift as usize].bitrev.as_ptr();
-    i = 0 as i32;
+    i = 0;
     while i < N4 {
         let mut rev: i32 = 0;
         let mut yr: f32 = 0.;
@@ -170,10 +168,10 @@ pub unsafe fn clt_mdct_backward_c(
         rev = *fresh8 as i32;
         yr = *xp2 * t[i as usize] + *xp1 * t[(N4 + i) as usize];
         yi = *xp1 * t[i as usize] - *xp2 * t[(N4 + i) as usize];
-        *yp.offset((2 as i32 * rev + 1 as i32) as isize) = yr;
-        *yp.offset((2 as i32 * rev) as isize) = yi;
-        xp1 = xp1.offset((2 as i32 * stride) as isize);
-        xp2 = xp2.offset(-((2 as i32 * stride) as isize));
+        *yp.offset((2 * rev + 1) as isize) = yr;
+        *yp.offset((2 * rev) as isize) = yi;
+        xp1 = xp1.offset((2 * stride) as isize);
+        xp2 = xp2.offset(-((2 * stride) as isize));
         i += 1;
     }
     opus_fft_impl(
@@ -183,46 +181,46 @@ pub unsafe fn clt_mdct_backward_c(
             l.kfft[shift as usize].nfft,
         ),
     );
-    let mut yp0: *mut f32 = out.offset((overlap >> 1 as i32) as isize);
+    let mut yp0: *mut f32 = out.offset((overlap >> 1) as isize);
     let mut yp1: *mut f32 = out
-        .offset((overlap >> 1 as i32) as isize)
+        .offset((overlap >> 1) as isize)
         .offset(N2 as isize)
-        .offset(-(2 as i32 as isize));
+        .offset(-(2 as isize));
     let t = trig;
-    i = 0 as i32;
-    while i < N4 + 1 as i32 >> 1 as i32 {
+    i = 0;
+    while i < N4 + 1 >> 1 {
         let mut re: f32 = 0.;
         let mut im: f32 = 0.;
         let mut yr_0: f32 = 0.;
         let mut yi_0: f32 = 0.;
         let mut t0: f32 = 0.;
         let mut t1: f32 = 0.;
-        re = *yp0.offset(1 as i32 as isize);
-        im = *yp0.offset(0 as i32 as isize);
+        re = *yp0.offset(1 as isize);
+        im = *yp0.offset(0 as isize);
         t0 = t[i as usize];
         t1 = t[(N4 + i) as usize];
         yr_0 = re * t0 + im * t1;
         yi_0 = re * t1 - im * t0;
-        re = *yp1.offset(1 as i32 as isize);
-        im = *yp1.offset(0 as i32 as isize);
-        *yp0.offset(0 as i32 as isize) = yr_0;
-        *yp1.offset(1 as i32 as isize) = yi_0;
+        re = *yp1.offset(1 as isize);
+        im = *yp1.offset(0 as isize);
+        *yp0.offset(0 as isize) = yr_0;
+        *yp1.offset(1 as isize) = yi_0;
         t0 = t[(N4 - i - 1) as usize];
         t1 = t[(N2 - i - 1) as usize];
         yr_0 = re * t0 + im * t1;
         yi_0 = re * t1 - im * t0;
-        *yp1.offset(0 as i32 as isize) = yr_0;
-        *yp0.offset(1 as i32 as isize) = yi_0;
-        yp0 = yp0.offset(2 as i32 as isize);
-        yp1 = yp1.offset(-(2 as i32 as isize));
+        *yp1.offset(0 as isize) = yr_0;
+        *yp0.offset(1 as isize) = yi_0;
+        yp0 = yp0.offset(2 as isize);
+        yp1 = yp1.offset(-(2 as isize));
         i += 1;
     }
-    let mut xp1_0: *mut f32 = out.offset(overlap as isize).offset(-(1 as i32 as isize));
+    let mut xp1_0: *mut f32 = out.offset(overlap as isize).offset(-(1 as isize));
     let mut yp1_0: *mut f32 = out;
     let mut wp1: *const opus_val16 = window;
-    let mut wp2: *const opus_val16 = window.offset(overlap as isize).offset(-(1 as i32 as isize));
-    i = 0 as i32;
-    while i < overlap / 2 as i32 {
+    let mut wp2: *const opus_val16 = window.offset(overlap as isize).offset(-(1 as isize));
+    i = 0;
+    while i < overlap / 2 {
         let mut x1: f32 = 0.;
         let mut x2: f32 = 0.;
         x1 = *xp1_0;
